Flying Bird mini game
When you were young, do you ever play a game while you have to make a bird fly without touching the pipe by pressing the screen make the bird fly? This is the same game but a bit similar and lack of some feature but also define as one of my fun project so feel free to check it out.